# [jvm-packages] Clean the way deterministic partitioning is computed