We propose to only use the rowHashCode to compute the partitionKey, adding the FeatureValue hashCode might not bring a better load balancing and is makingthe computation slower. MurmurHash3 is known as being quite uniformly distributed and should suffice here.