@Totoketchup 2020-08-19
Yes it is both fixing the problem with Double.NaN and making the deteministic repartitioning faster.Linked to this remark I made: https://github.com/dmlc/xgboost/pull/5996#issuecomment-671256383