@CodingCat 2020-08-19
row.hashCode returns the hashCode purely based on the murmur3, we have seen cases that some partition got empty in prod, 

the code here is to add another layer of hash to murmur3, i.e. even murmur3 is the same for two rows, if their particular feature has different value, they will be assigned to different partition 

>  is makingthe computation slower

could you quantify this? how this becomes a bottleneck?