@Totoketchup 2020-08-20
Thanks for your answer ;) 

> we have seen cases that some partition got empty in prod.
This looks more like a bug (like we had with NaN values) than something coming from MurMurHash3, no ? (but I might be wrong). By any chance, do still have the dataset ? I could run some analyses to see the load balancing on it. 

What I'll do is that I'll run some load balancing analyses on the test I added last review first to see if we see any big changes in the distribution of rows in the partitions between MurmurHash3 and MurmurHash3 + Feature Value. WDYT ?

> could you quantify this? how this becomes a bottleneck?

For the speed-up I did not run any analyses indeed but I just 'guessed' that accessing the FeatureVector to retrieve a new value, and then computing its hash is adding some computational time. But maybe the impact is really negligible yes. I'll use the test I added to do some benches on 1M lines. This is not really a bottleneck strictly speaking, but if we can gain some time and not lose any load balancing, this is nice.