@Totoketchup 2020-08-23
I ran some benchmarks using the test I previously added on `val trainingDF = buildDataFrame(Classification.train)` with 1000 runs and there is a slight speedup in average. But this might be negligible.

```
Current:
  // Max   227,483,731 ns
  // Min    11,876,080 ns
  // Avg    23,015,432 ns

PR:
  // Max    62,906,984 ns
  // Min    11,879,247 ns
  // Avg    18,362,924 ns
```