@Totoketchup 2020-08-23
The thing is that for MurmurHash3 we must look at how uniform is the distribution of hashes and not at collisions. For sure we have high level of collision because of the partitioning size and the modulus operator. Even though 2 rows are different with MurmurHash3, they could collide with the modulus operator. Looking at https://stackoverflow.com/questions/11899616/murmurhash-what-is-it, it confirms that MurmurHash has quite a good uniform distribution, so applying the modulus operator on it would still give a good distribution in my opinion.
The scenario you describe where an entire partition is entirely empty is almost impossible.