@CodingCat 2020-08-23
> The scenario you describe where an entire partition is entirely empty is almost impossible.

did I misunderstand anything here? You mean what I saw in prod is fake?