@Totoketchup 2020-08-23
That's not what I'm saying, deeply sorry if you understood it like this. If you're telling that this happened to you, of course I believe you. What I'm wondering, from what I read about MurmurHash3 (its uniform distribution and its good avalanche effect) what are the number of rows / columns you have in your prod dataset and the number of workers you used to see such a behavior ?

It's not that I'm not believing you, I am, what I think that what might have happened to you is maybe like what we had with Double.NaN where 50% of our dataset was in only one partition.