@Totoketchup 2020-08-27
I'd like to go a bit more in the anlyses of such a scenario:

Let's imagine the scenario where there is indeed a partiition completely empty after having used the repartitioning and let's go backward from this situation.

Let's say we have `N` rows and `P` partitions, and name the faulty parition `P_x`.

To know where a Row R_i is going, we compute `hash(R_i) % P`.
Now, what we have in this case is that for all rows `R_i: hash(R_i) % P != P_x`.  Which means that in this case the hash function is **never** returning a number where the remainder is never `x` but all other numbers in [1, P]. 
From this, we can see that this would be a huge flaw in MurmurHash3, right ?
Another thing is that by just changing P, then the whole behavior would change and your empty partition is not existing anymore.

What I mean by this analysis is that what you are observing might not be due to the hash function itself.
Do you agree with me ?